+++ src/main.h
@@ -1,0 +1,25 @@
+//#pragma once
+#ifndef MAIN_HEADER_23424
+#define MAIN_HEADER_23424
+
+#include <Arduino.h>
+
+
+
+
+
+
+
+
+
+
+#if CONFIG_FREERTOS_UNICORE
+#define ARDUINO_RUNNING_CORE 0
+#else
+#define ARDUINO_RUNNING_CORE 1
+#endif
+
+
+
+
+#endif
